update page sturcture
