+++ ymir/web/src/pages/project/detail.js
@@ -57,7 +57,7 @@ function ProjectDetail(func) {
     <div>
       <Breadcrumbs />
       <div className={s.header}>
-      {project.enableIteration ? <Detail project={project} iterations={iterations} fresh={fresh} /> : <NoIterationDetail project={project} />}
+        <Button onClick={() => history.push(`/home/project/${id}/iterations`)}>Go Iterations</Button>
       </div>
       <Space className="actions">
         <Button type="primary" onClick={add}><ImportIcon /> {t("dataset.import.label")}</Button>
