+++ ymir/web/src/pages/project/iterations/index.less
@@ -1,0 +1,38 @@
+.list {
+  .extraTag {
+    position: absolute;
+    right: 0;
+    top: 2px;
+    font-size: 12px;
+    line-height: 12px;
+    color: #fff;
+    text-align: right;
+    .negative, .positive, .neutral {
+      display: inline-block;
+      padding: 2px 4px;
+      border-radius: 4px;
+      &::after {
+        display: inline-block;
+        margin-left: 4px;
+      }
+    }
+    .negative {
+      background-color: darkred;
+      &::after {
+        content: '\2193';
+      }
+    }
+    .positive {
+      &::after {
+        content: '\2191';
+      }
+      background-color: darkgreen;
+    }
+    .neutral {
+      background-color: darkgray;
+    }
+  }
+}
+.orange {
+  color: orange;
+}
